feat: ProductDetailMain 추상화수준 통일
