+++ src/entities/product/ui/index.ts
@@ -1,4 +1,4 @@
 export * from "./ProductCardInList";
 export * from "./ProductDetailColorSelect";
 export * from "./ProductDetailSizeStockSelect";
-export * from "./ProductColorSizeSelects";
+export * from "./ProductHeart";
