+++ src/widgets/product/ui/index.ts
@@ -1,4 +1,8 @@
-export * from './ProductDetailCarousel'
-export * from './ProductDetailMain'
-export * from './ProductDescription'
-export * from './ProductDetailMainLayout'
+export * from "./ProductDetailCarousel";
+export * from "./ProductDetailMain";
+export * from "./ProductDescription";
+export * from "./ProductDetailMainLayout";
+export * from "./ProductColorSizeSelects";
+export * from "./ProductDiscountPrice";
+export * from "./SelectedProductsTable";
+export * from "./ProductDetailTotalPrice";
